Ingest workflow working
Ingest has been split into two major workflow (via celery canvases) to
be less of a stove pipe solution. Currently, only images/KRT are
supported using the new method. The old ingest tasks are left in and
wired, but may be buggy due to the removal of extra directories in
ingest. This can be fixed as each one is ported to the new ingest
system.
-Got celery canvases working
-Removed multiple directories from ingest, due to HTML 5.0 limitations
-Added Cleanup task to ingest workflow
-Added original_filename to Image model for ingest metadata matching,
etc...
-Various code refactors and pep8 updates